JL8-27 fix : sideBar 가려지는 현상 수정
